Updated onConnection related to fix multiple handler issue
Since we fixed multiple handler issue by removing the extra io.on('connection') event from our socketConnect.js file, we are moving the console log of the new connection here. Notice we do not need to call io.on connection as that is what is calling this function in the first place. The console log sits at the top of the onConnection function alone.