script: Added functionality for toggle pass btn
Using the same kind of global interval I used for the error divs, I created a timer and on subsequent presses I cancel the current one and start a new timer. Prevents having tons of timers started and going off. I wanted the field to reset after 10 seconds, and I also made it so the effect is shared between the roomListOverlay and the join/createRoom UI.